Upload project notebook for Bank Customer Churn Prediction using Neural Network
Added the final Jupyter Notebook for the Bank Customer Churn Prediction project. This notebook includes data preprocessing, neural network model design, training, evaluation, and key findings. The model aims to classify whether a customer is likely to leave the bank in the next 6 months based on behavioral and demographic features.